Event buffer return values and exceptions
The upcoming event buffer functionality requires
the Pusher::trigger function to expose the event_id
of each event triggered on a channel. In order to
do this the success return value has to be changed
from `true` to an object that allows the event_id
values to be inspected.

As part of this work the library is also updated
to throw exceptions when any unexpected response
is returned from the Pusher HTTP API.

Additional details:

* `get` and `post` now use internal `request` function
to share common functionality.
* `get` and `post` return a more “raw” response
that exposes a) response status code b) response body