Update Pusher.php
The uri is badly composed, when url path is defined in config. Base uri for Guzzle is concatenated with config path and therefore, it does not work. Also, in custom websocket server, I need to override the socket id validation to the custom rule, so please set it as protected.